Initial refactoring: - Break original method into services
